zmatrices working for all fastfolders
