+++ bgmol/tests/conftest.py
@@ -2,8 +2,10 @@
 
 import pytest
 import os
+from itertools import product
 from bgmol.datasets import Ala2Implicit1000
 from bgmol.systems import ChignolinC22Implicit
+from bgmol.systems.fastfolders import FastFolder, FAST_FOLDER_NAMES
 from bgmol import systems, api
 
 
@@ -37,6 +39,19 @@ def chignolin(tmpdir_factory):
     return ChignolinC22Implicit(root=tmpdir_factory.mktemp("chignolin_system"))
 
 
+@pytest.fixture(scope="session", params=product(
+    FAST_FOLDER_NAMES,
+    (False, True),
+    ("charmm36m", ["amber99sbildn.xml", "tip3p.xml"])
+))
+def fastfolder_system(request, tmpdir_factory):
+    protein, solvated, forcefield = request.param
+    tmpdir = tmpdir_factory.mktemp(protein)
+    if forcefield != "charmm36m" and protein in ["ntl9", "villin"]:
+        pytest.skip("Nonstandard residues.")
+    yield FastFolder(protein, download=True, solvated=solvated, forcefield=forcefield, root=str(tmpdir))
+
+
 # skipping slow tests by default
 # ==============================
 def pytest_addoption(parser):
@@ -57,3 +72,6 @@ def pytest_collection_modifyitems(config, items):
     for item in items:
         if "slow" in item.keywords:
             item.add_marker(skip_slow)
+
+
+
